LPS-23595 Classes that cover cas AttributePrincipal
This should be needed so plugins can have access to LiferayCasPrincipal,
which is placed in portal-service. cas-client does not provide different
jars for api and impl, so we could not deploy AttributePrincipal in
portal-service. LiferayCasPrincipal is used in replacement (at least at
the moment) of AttributePrincipal and the portlets and plugins may use
this class to get tickets for remote services.