LPS-47465 Adapt to new interface
Now the ServiceTrackerMapImpl will ask the service tracker to track the
service reference and service tuples. This implies the
ServiceTrackerMapImpl MUST call just one time getService() for each
tuple and will handle the tracker service along with the service
reference to the buckets using the tuple.

Furthermore if no key is emitted by the mapper then a null will be
returned so that the service tracker does not track that tuple (since it
would not be accesible by any key).

The API to the user has been left unchanged.