LPS-91324 fix the javax.mvc-api dependency
This API depends on JAX-RS, which is already provide by JAX-RS Whiteboard module. So we cannot deploy javax.mvc-api as static module since it won't resolve. This commit deploys it from bean-portlet-extension-api which is the only common bundle to bean-portlet-cdi-extension and bean-portlet-spring-extension that need it; which is also why I made it a runtime dependency.